requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
questionary==2.0.1
rich==13.4.2
redis==4.5.5